Split the server.rs file into several
