+++ src/lib.rs
@@ -32,5 +32,9 @@ pub use addr::Addr;
 pub mod result;
 pub mod addr;
 pub mod tag;
+mod health;
+mod graph;
+mod state;
+mod node;
 mod broadcast;
 mod server;
