+++ src/server.rs
@@ -3,7 +3,6 @@
 use std::collections::hashmap::HashSet;
 use std::comm::{Receiver, Sender};
 use std::rc::{Rc, Weak};
-use std::sync::Arc;
 use uuid::Uuid;
 use std::io::timer::Timer;
 use std::time::duration::Duration;
@@ -12,18 +11,9 @@ use addr::Addr;
 use broadcast::Broadcast;
 use result::{GossipResult, GossipError};
 
-/// A health represents the current state of the cluster. This will be extremely useful
-/// to ping the health of a cluster and determine the high-level status of it.
-///
-/// Green = Perfect state.
-/// Yellow = Nodes are failing, but the cluster is still operational.
-/// Red = Not good. Cluster might be completely dead.
-#[deriving(Show, PartialEq, Clone)]
-pub enum Health {
-    Green,
-    Yellow,
-    Red
-}
+use node::Node;
+use health::{Health};
+use state::State;
 
 #[deriving(Show, PartialEq)]
 pub enum ShutdownReason {
@@ -43,62 +33,6 @@ pub enum ServerMsg {
     /// Kill a specific node in the cluster. This is a state change rather than a gossip. This will
     /// remove a specific node from the cluster.
     KillNode(Node)
-}
-
-#[deriving(PartialEq, Clone)]
-pub struct State {
-    eager: HashSet<Node>,
-    lazy: HashSet<Node>,
-    health: Health,
-    broadcasts: Vec<Broadcast>,
-    graph: Graph
-}
-
-impl State {
-    /// Create a new default State that starts a new cluster in a
-    /// Yellow state.
-    pub fn new() -> State {
-        State {
-            eager: HashSet::new(),
-            lazy: HashSet::new(),
-            health: Yellow,
-            broadcasts: Vec::new(),
-            graph: Graph::new()
-        }
-    }
-}
-
-/// The graph representation of our communication model. The most ideal representation
-/// would be a spanning tree, however, that's not always possible because of the
-/// highly-available properties of our distributed system. A spanning tree would essentially
-/// prove to be the most minimal set of communication points possible to achieve
-/// the successful distribution of our broadcasts.
-///
-/// We'll have to periodically compute if the graph is a spanning tree or not.
-#[deriving(PartialEq, Clone)]
-pub struct Graph {
-    /// We group the graph by vertices so we can easily fetch all the edges of a
-    /// particular vertex.
-    vertices: HashSet<Vertex>,
-    /// Is the tree in spanning mode? This should ensure that we are
-    /// in an optimized-mode.
-    spanning: bool
-}
-
-impl Graph {
-    /// Create a new graph with an empty set and a default of spanning: false
-    pub fn new() -> Graph {
-        Graph {
-            vertices: HashSet::new(),
-            spanning: false
-        }
-    }
-}
-
-#[deriving(Eq, PartialEq, Hash, Clone)]
-pub struct Vertex {
-    server: Node,
-    edges: Vec<Rc<Vertex>>
 }
 
 pub struct ServerTask {
@@ -262,50 +196,10 @@ impl Server {
     }
 }
 
-/// A node is a server within the cluster without any state associated with it. We
-/// only keep state and things like channels for the current server, not other ones in the cluster.
-/// A node handles the concept of a node within the cluster that we need to interface with.
-///
-/// Each server holds enough metadata to work within the cluster, such as all the current members
-/// of the cluster. We only need a small number of details for those servers, however, so we'd use
-/// a Node instead of the Server record.
-#[deriving(Show, Eq, PartialEq, Hash, Clone)]
-pub struct Node {
-    id: Uuid,
-    addr: Addr
-}
-
-impl Node {
-    /// Create a new node given an ip address and a port. This does not actually
-    /// connect to that node or anything. They are simply identifiers.
-    /// The transport handles
-    pub fn new(ip: &str, port: u16) -> Node {
-        Node {
-            id: Uuid::new_v4(),
-            addr: Addr::new(ip, port)
-        }
-    }
-}
 
 #[cfg(test)]
-mod test {
+mod tests {
     use super::*;
-
-    #[test]
-    fn default_graph() {
-        let g = Graph::new();
-        assert_eq!(g.spanning, false);
-        assert_eq!(g.vertices.len(), 0);
-    }
-
-    #[test]
-    fn default_state() {
-        let s = State::new();
-        assert_eq!(s.eager.len(), 0);
-        assert_eq!(s.lazy.len(), 0);
-        assert_eq!(s.broadcasts.len(), 0);
-        assert_eq!(s.health, Yellow);
-    }
 
     #[test]
     fn default_server() {
